Update play services verification dialog check
Changed to not display error dialog if play services are missing or disabled.
Most of the time this means that user doesn't want or can't run Play Services.
We don't want to bother those users with a dialog. 
Alternative solution would be to show a dialog with 'never again' checkbox, 
but it seems like too much work for a small gain.